Fix build on Red Hat 6
The RH compiler is a bit more strict about warnings for signed
variables, this just adds some casts to eliminate the warnings.  It's
not so clean to require these casts, but they're just in some
assert()s so I guess it's OK for now.

+++ src/pconfigure/main.c
@@ -202,7 +202,7 @@ int parse_line(const char *line, char *left, char *op, char *right)
     /* Splits into 3 parts */
     while ((line[line_cur] != '\0') && !isspace(line[line_cur]))
     {
-        assert(left_cur < MAX_LINE_SIZE);
+        assert(left_cur < (int)MAX_LINE_SIZE);
         left[left_cur] = line[line_cur];
         left_cur++;
         line_cur++;
@@ -213,7 +213,7 @@ int parse_line(const char *line, char *left, char *op, char *right)
 
     while ((line[line_cur] != '\0') && !isspace(line[line_cur]))
     {
-        assert(op_cur < MAX_LINE_SIZE);
+        assert(op_cur < (int)MAX_LINE_SIZE);
         op[op_cur] = line[line_cur];
         op_cur++;
         line_cur++;
@@ -224,7 +224,7 @@ int parse_line(const char *line, char *left, char *op, char *right)
 
     while (line[line_cur] != '\0')
     {
-        assert(right_cur < MAX_LINE_SIZE);
+        assert(right_cur < (int)MAX_LINE_SIZE);
         right[right_cur] = line[line_cur];
         right_cur++;
         line_cur++;
